SCALE-SEAM (87) — the map described the tail, not the file
Thirteen methods out of client.ts (1,131 -> 1,015) into eight existing
mixins: material palette to elements.ts, test fit to entitlements.ts beside
feasibility, property to proforma.ts, rebar BBS to estimate.ts, and five
singles to modules.ts, procurement.ts, schedule.ts and evm.ts. Five types
moved to types.ts with them.

THE FINDING. Deriving the whole population instead of trusting the banner:
130 methods in client.ts, 4 below it, 126 ABOVE it. Those 126 were never
inside the RACI or CX-1 banners, so no map has ever covered them. The
UNFILED map described the TAIL of this file, not the file, and every "N
remain" made against it was scoped to a list the same slices had authored.
The banner is renamed UNFILED -> STAYING, reduced to the four global methods
it actually decides, and now says so in its own text.

THE PREDICATE WAS REPLACED TWICE. test_roadmap_status held SCALE-SEAM open
on "client.ts > 1200" — a threshold proxy that decayed and went red at (85).
Its replacement checked for the banner string, and that was ALSO a proxy: it
would have declared SCALE-SEAM done with 126 methods outstanding, and I
described it as a fix in a commit, a PR body and to the user before noticing.
It now counts methods not on the declared keep-list, mutation-checked four
ways (baseline 126 OPEN, banner renamed still OPEN, keep-list deleted 130 so
it fails safe, simulated completion 0 DONE).

TWO CORRECTIONS TO EARLIER SLICES. pxSummary is the programme-level rollup
that belonged with the rollup/detail split evm.ts made at (9); modules.ts
carried a doc comment orphaned from complianceExpiring, swept in without its
method. Both reunited.

RATCHET HISTORY CHAIN GATED. test_file_sizes records each slice as
(before -> after) beside the pin; at (86) a narrative said (1173 -> 1132)
against a pin of 1131. Consecutive pairs must now link, and an unrecorded
shrink fails while growth does not. Four tries, all four failures found by
running it — \d+ skipped underscores, a closing-paren requirement invented a
false break, strict equality flagged legitimate growth, and the finished
checks printed FAIL but exited 0 because they sat below the exit block.
viewer/app.ts carries a named exemption for its documented 2571/2570 slack.
